Ultimate challenge 3
first we encountered an error. We addressed this error by creating an empty hello_world.rb file. I uncommented the skip part in test file and then ran again. Then we got no errors. Then for the hello_world_test.rb, we did the similar steps. We first got error, as the initial step of TDD and we created new file and we had no errors. We did the same thing until 05-grains.